requirements.txt
lxml==4.4.1
requests==2.22.0
requests_futures==1.0.0
Pillow==6.1.0
beautifulsoup4==4.8.0
